discovered a bug in the queue
In the queue's "add function" function, the loop did not iterate through
the circular list to find an empty node.  It counted through the number
of nodes, but it only checked the current node's function pointer.

Added in an "else" statement to advance the node pointer if the current
node is used.